+++ src/main/java/com/programmers/springbootbasic/presentation/view/Input.java
@@ -1,4 +1,4 @@
-package com.programmers.springbootbasic.common;
+package com.programmers.springbootbasic.presentation.view;
 
 import java.io.IOException;
